+++ src/server/routers/lambda/index.ts
@@ -28,6 +28,7 @@ import { sessionRouter } from './session';
 import { sessionGroupRouter } from './sessionGroup';
 import { threadRouter } from './thread';
 import { topicRouter } from './topic';
+import { uploadRouter } from './upload';
 import { userRouter } from './user';
 
 export const lambdaRouter = router({
@@ -57,6 +58,7 @@ export const lambdaRouter = router({
   sessionGroup: sessionGroupRouter,
   thread: threadRouter,
   topic: topicRouter,
+  upload: uploadRouter,
   user: userRouter,
 });